Made a seperate file for ProjectCard component
